html/browsers/browsing-the-web/back-forward-cache/storage-events.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
// https://github.com/whatwg/storage/issues/119#issuecomment-785771108
promise_test(async t => {
  const pageA = new RemoteContext(token());
  const pageB = new RemoteContext(token());
  const pageC = new RemoteContext(token());

  const urlA = executorPath + pageA.context_id + '&events=pagehide,pageshow,load';
  const urlB = originCrossSite + executorPath + pageB.context_id;
  const urlC = executorPath + pageC.context_id + '&events=pagehide,pageshow,load';

  // localStorage keys set while pageA is active/inactive.
  const key1 = token();
  const key2 = token();

  const startRecordingStorageEvent = (key1, key2) => {
    window.addEventListener('storage', e => {
      if (e.key === key1) {
        recordEvent('storage1');
      }
      if (e.key === key2) {
        recordEvent('storage2');
      }
    });
  };

  window.open(urlA, '_blank', 'noopener');
  await pageA.execute_script(waitForPageShow);
  await pageA.execute_script(startRecordingStorageEvent, [key1, key2]);

  // Window C is used to check that failures are BFCache-related, not
  // underlying issues around storage events.
  window.open(urlC, '_blank');
  await pageC.execute_script(waitForPageShow);
  await pageC.execute_script(startRecordingStorageEvent, [key1, key2]);

  // Navigate A to B.
  await pageA.execute_script((url) => {
    prepareNavigation(() => {
      location.href = url;
    });
  }, [urlB]);

  // Wait for navigation.
  await pageB.execute_script(waitForPageShow);

  // Update `key1` while pageA is inactive.
  localStorage.setItem(key1, 'value');

  await pageB.execute_script(
    () => {
      prepareNavigation(() => { history.back(); });
    }
  );

  await pageA.execute_script(waitForPageShow);
  await assert_bfcached(pageA);

  // Update `key2` after pageA becomes active.
  localStorage.setItem(key2, 'value');

  assert_array_equals(
    await pageC.execute_script(() => getRecordedEvents()),
    [
      'window.load',
      'window.pageshow',
      'storage1',
      'storage2',
    ],
    'pageC should receive storage events');

  assert_array_equals(
    await pageA.execute_script(() => getRecordedEvents()),
    [
      'window.load',
      'window.pageshow',
      'window.pagehide.persisted',
      'window.pageshow.persisted',
      'storage1',
      'storage2',
    ],
    'pageA should receive storage events after it becomes active');

}, 'Storage events should be fired for BFCached pages after becoming active');
</script>
